model: Rename the `util` package to `utils`
In order to make the naming consistent with other utility packages
within ORT.
