+++ model/src/main/kotlin/utils/FindingsMatcher.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package com.here.ort.model.util
+package com.here.ort.model.utils
 
 import com.here.ort.model.CopyrightFinding
 import com.here.ort.model.CopyrightFindings
